+++ cpp/core/coordinator/include/basis/core/coordinator.h
@@ -3,10 +3,13 @@
 #include <cstdint>
 #include <list>
 
+#include <transport.pb.h>
+
 #include <basis/plugins/serialization/protobuf.h>
 #include <basis/plugins/transport/tcp.h>
 
-#include <transport.pb.h>
+
+#include "coordinator_default_port.h"
 
 /**
  * This is a bit annoying. We have to create a custom transport for publisher info as both the coordinator socket needs
@@ -34,18 +37,20 @@
  *
  * TODO: it might be simpler to just write this as a protobuf grpc server :)
  */
-constexpr uint16_t BASIS_PUBLISH_INFO_PORT = 1492;
-
-// constexpr char BASIS_PUBLISHER_INFO_TOPIC[] = "/basis/publisher_info";
 
 namespace basis::core::transport {
 /**
- *
+ * A utility class for communicating the topic network state between TransportManagers (typically one per process)
+ * 
  * Implemented single threaded for safety - latency isn't a huge concern here.
  *
  * @todo track publishers and add a delta message
+ * @todo on connection, send the last message again (can we just implement latching?) 
  */
 class Coordinator {
+  /**
+   * An internal connection to a TransportManager
+   */
   struct Connection : public basis::plugins::transport::TcpSender {
     Connection(core::networking::TcpSocket socket) : TcpSender(std::move(socket)) { socket.SetNonblocking(); }
 
@@ -57,147 +62,41 @@ class Coordinator {
   };
 
 public:
-  static std::optional<Coordinator> Create(uint16_t port = BASIS_PUBLISH_INFO_PORT) {
-    // todo: maybe return the listen socket error type?
-    auto maybe_listen_socket = networking::TcpListenSocket::Create(port);
-    if (!maybe_listen_socket) {
-        spdlog::error("Coordinator: Unable to create listen socket on port {}");
-      return {};
-    }
-
-    return Coordinator(std::move(maybe_listen_socket.value()));
-  }
+  /**
+   * Creates a coordinator.
+   *
+   * @param port overridable - by default BASIS_PUBLISH_INFO_PORT
+   * @return a Coordinator, or std::nullopt if creation fails (usually due to the port already being taken)
+   */
+  static std::optional<Coordinator> Create(uint16_t port = BASIS_PUBLISH_INFO_PORT);
 
   Coordinator(networking::TcpListenSocket listen_socket) : listen_socket(std::move(listen_socket)) {}
 
-  proto::NetworkInfo GenerateNetworkInfo() {
-    proto::NetworkInfo out;
+  /**
+   * Gathers the publisher information from each client and packages it up into one message.
+   */
+  proto::NetworkInfo GenerateNetworkInfo();
 
-    for (const auto &client : clients) {
-      if (client.info) {
-        for (auto &publisher : client.info->publishers()) {
-          *(*out.mutable_publishers_by_topic())[publisher.topic()].add_publishers() = publisher;
-        }
-      }
-    }
+  /**
+   * Should be called on a regular basis, from the main thread.
+   *
+   * Accept new clients
+   * Compiles data about the subscription networks
+   * Sends the data to each client
+   *
+   */
+  void Update();
 
-    return out;
-  }
-
-  void Update() {
-    // accept new clients
-    // select on all clients
-    // compile new data on publishers
-
-    //
-    while (auto maybe_socket = listen_socket.Accept(0)) {
-      clients.emplace_back(std::move(maybe_socket.value()));
-    }
-
-    for (auto it = clients.begin(); it != clients.end();) {
-      auto &client = *it;
-      switch (client.ReceiveMessage(client.in_progress_packet)) {
-      case plugins::transport::TcpConnection::ReceiveStatus::DONE: {
-        auto complete = client.in_progress_packet.GetCompletedMessage();
-        client.info = basis::DeserializeFromSpan<proto::TransportManagerInfo>(complete->GetPayload());
-        spdlog::info("Got completed message {}", client.info->DebugString());
-
-        // convert to proto::PublisherInfo
-        // fallthrough
-      }
-      case plugins::transport::TcpConnection::ReceiveStatus::DOWNLOADING: {
-        // No work to be done
-        ++it;
-        break;
-      }
-      case plugins::transport::TcpConnection::ReceiveStatus::ERROR: {
-        spdlog::error("Client connection error after bytes {} - got error {} {}",
-                      client.in_progress_packet.GetCurrentProgress(), errno, strerror(errno));
-        it = clients.erase(it);
-        break;
-              }
-      case plugins::transport::TcpConnection::ReceiveStatus::DISCONNECTED: {
-                spdlog::error("Client connection disconnect after {} bytes",
-                      client.in_progress_packet.GetCurrentProgress());
-        // TODO: we can do a fast delete here instead, swapping with .last()
-        it = clients.erase(it);
-        break;
-      }
-      }
-    }
-
-    proto::NetworkInfo info = GenerateNetworkInfo();
-
-    using Serializer = SerializationHandler<proto::TransportManagerInfo>::type;
-    const size_t size = Serializer::GetSerializedSize(info);
-
-    auto shared_message = std::make_shared<basis::core::transport::MessagePacket>(
-        basis::core::transport::MessageHeader::DataType::MESSAGE, size);
-    Serializer::SerializeToSpan(info, shared_message->GetMutablePayload());
-
-    for (auto &client : clients) {
-      client.SendMessage(shared_message);
-    }
-  }
-
+protected:
+  /**
+   * The TCP listen socket.
+   */
   core::networking::TcpListenSocket listen_socket;
 
+  /**
+   * The clients we should send information to. On disconnection, will be removed by Update()
+   */
   std::list<Connection> clients;
 };
 
-class CoordinatorConnector : basis::plugins::transport::TcpSender {
-  using TcpSender::TcpSender;
-
-public:
-  static std::unique_ptr<CoordinatorConnector> Create(uint16_t port = BASIS_PUBLISH_INFO_PORT) {
-    auto maybe_socket = networking::TcpSocket::Connect("127.0.0.1", port);
-    if (maybe_socket) {
-      return std::make_unique<CoordinatorConnector>(std::move(*maybe_socket));
-    }
-    return {};
-  }
-  // TODO: this logic is somewhat duplicated with TcpSender - need to refactor TcpSender to be able to track progress of
-  // the current send
-  void SendTransportManagerInfo(const proto::TransportManagerInfo &info) {
-    using Serializer = SerializationHandler<proto::TransportManagerInfo>::type;
-    const size_t size = Serializer::GetSerializedSize(info);
-
-    auto shared_message = std::make_shared<basis::core::transport::MessagePacket>(
-        basis::core::transport::MessageHeader::DataType::MESSAGE, size);
-    Serializer::SerializeToSpan(info, shared_message->GetMutablePayload());
-
-    SendMessage(shared_message);
-  }
-
-  void Update() {
-    // todo: just put this on tcpconnection??
-    switch (ReceiveMessage(in_progress_packet)) {
-      case plugins::transport::TcpConnection::ReceiveStatus::DONE: {
-        auto complete = in_progress_packet.GetCompletedMessage();
-        last_network_info = basis::DeserializeFromSpan<proto::NetworkInfo>(complete->GetPayload());
-        spdlog::info("Got network info message {}", last_network_info->DebugString());
-
-        // convert to proto::PublisherInfo
-        // fallthrough
-      }
-      case plugins::transport::TcpConnection::ReceiveStatus::DOWNLOADING: {
-        break;
-      }
-      case plugins::transport::TcpConnection::ReceiveStatus::ERROR: {
-        spdlog::error("connection error after bytes {} - got error {} {}",
-                      in_progress_packet.GetCurrentProgress(), errno, strerror(errno));
-        // fallthrough
-      }
-      case plugins::transport::TcpConnection::ReceiveStatus::DISCONNECTED: {
-        // TODO: we can do a fast delete here instead, swapping with .last()
-        break;
-      }
-      }
-  }
-
-  std::unique_ptr<proto::NetworkInfo> last_network_info;
-    
-  IncompleteMessagePacket in_progress_packet;
-
-};
 } // namespace basis::core::transport
