+++ cpp/core/coordinator/include/basis/core/coordinator_default_port.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <cstdint>
+
+constexpr uint16_t BASIS_PUBLISH_INFO_PORT = 1492;
